+++ src/App.tsx
@@ -2,7 +2,7 @@
 import { Authenticator } from '@aws-amplify/ui-react';
 import '@aws-amplify/ui-react/styles.css';
 
-import AuthenticatedContent from './pages/AuthenticatedContent';
+import AuthenticatedShell from './pages/AuthenticatedShell';
 
 function AuthHeader() {
   return (
@@ -24,7 +24,7 @@ export default function App() {
   return (
     <Authenticator components={{ Header: AuthHeader }}>
       {/* At this point we’re authenticated; no null return means no TS error */}
-      <AuthenticatedContent />
+      <AuthenticatedShell />
     </Authenticator>
   );
 }
